requirements-dev.txt
# Development Dependencies for myHelperBuddy

# Django Debug Toolbar for performance monitoring
django-debug-toolbar==4.2.0

# Testing framework
pytest==7.4.3
pytest-django==4.7.0

# Code formatting
black==23.12.0

# Linting
flake8==6.1.0

# Type checking
mypy==1.7.1
django-stubs==4.2.7

# Coverage reporting
coverage==7.3.3
pytest-cov==4.1.0

# Factory for test data
factory-boy==3.3.0
